+++ src/regular.rs
@@ -6,10 +6,11 @@ use sdl3::{pixels::{Color, PixelFormat}, rect::Rect, render::{Canvas, TextureCre
 
 
 /// Renders text without sub-pixel rendering (a bit faster and easier to use, but looks a bit pixelated)
-pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f32, x: i32, y: i32, foreground: Color, canvas: &mut Canvas<Window>, texture_creator: &'a TextureCreator<WindowContext>, text_cache: &mut TextCache<'a, F>) -> Result<(), RenderTextError> {
+pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f32, x: i32, y: i32, h_align: HAlign, v_align: VAlign, foreground: Color, canvas: &mut Canvas<Window>, texture_creator: &'a TextureCreator<WindowContext>, text_cache: &mut TextCache<'a, F>) -> Result<(), RenderTextError> {
 	let text = text.as_ref();
 	if text.is_empty() {return Ok(());}
 	let font = text_cache.font.as_scaled(PxScale::from(100.0));
+	
 	// convert chars to glyphs & rasterize uncached glyphs
 	let mut glyphs = Vec::with_capacity(text.len());
 	let new_textures = Mutex::new(vec!());
@@ -28,6 +29,7 @@ pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f
 			glyphs.push((c, glyph));
 		}
 	});
+	
 	// upload new glyph textures to gpu
 	for texture_data in new_textures.into_inner().unwrap() {
 		let Some((c, pixels, width, height, x_offset, y_offset)) = texture_data else {continue;};
@@ -41,19 +43,35 @@ pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f
 		text_cache.map_regular.insert((c, foreground), (texture, width, height, x_offset, y_offset));
 	}
 	let font = text_cache.font.as_scaled(PxScale::from(size));
+	
+	// get text width & align properly
+	let mut width = 0.0;
+	let (c, first_glyph) = &glyphs[0];
+	width += font.h_advance(first_glyph.id);
+	width += size * EXTRA_CHAR_SPACING;
+	if c.is_whitespace() {width += size * EXTRA_WHITESPACE_SPACING;}
+	for [(_prev_c, prev_glyph), (c, glyph)] in glyphs.array_windows() {
+		width += font.kern(prev_glyph.id, glyph.id);
+		width += font.h_advance(glyph.id);
+		width += size * EXTRA_CHAR_SPACING;
+		if c.is_whitespace() {width += size * EXTRA_WHITESPACE_SPACING;}
+	}
+	width -= size * EXTRA_CHAR_SPACING;
+	let mut x = x as f32 + h_align.get_offset(width);
+	let y = y as f32 + v_align.get_offset(font.height());
+	
 	// render first char
-	let mut x = x as f32;
-	let y = y as f32 + font.height() * 0.75;
-	if let Some((c, glyph)) = glyphs.first() {
+	if let Some((c, first_glyph)) = glyphs.first() {
 		let texture_data = text_cache.map_regular.get(&(*c, foreground));
 		if let Some((texture, width, height, x_offset, y_offset)) = texture_data {
 			let dst = Rect::new((x - *x_offset * size / 100.0) as i32, (y - *y_offset * size / 100.0) as i32, (size * (*width as f32 / 100.0)) as u32, (size * (*height as f32 / 100.0)) as u32);
 			canvas.copy(texture, None, dst)?;
 		}
-		x += font.h_advance(glyph.id);
-		x += size as f32 * EXTRA_CHAR_SPACING;
-		if c.is_whitespace() {x += size as f32 * EXTRA_WHITESPACE_SPACING;}
+		x += font.h_advance(first_glyph.id);
+		x += size * EXTRA_CHAR_SPACING;
+		if c.is_whitespace() {x += size * EXTRA_WHITESPACE_SPACING;}
 	}
+	
 	// render remaining chars (with kerning)
 	for [(_prev_c, prev_glyph), (c, glyph)] in glyphs.array_windows() {
 		x += font.kern(prev_glyph.id, glyph.id);
@@ -63,9 +81,10 @@ pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f
 			canvas.copy(texture, None, dst)?;
 		}
 		x += font.h_advance(glyph.id);
-		x += size as f32 * EXTRA_CHAR_SPACING;
-		if c.is_whitespace() {x += size as f32 * EXTRA_WHITESPACE_SPACING;}
+		x += size * EXTRA_CHAR_SPACING;
+		if c.is_whitespace() {x += size * EXTRA_WHITESPACE_SPACING;}
 	}
+	
 	Ok(())
 }
 
@@ -73,7 +92,7 @@ pub fn render_text_regular<'a, F: ThreadSafeFont>(text: impl AsRef<str>, size: f
 
 fn rasterize_glyph_regular(glyph: Glyph, c: char, foreground: Color, font: &PxScaleFont<&impl ThreadSafeFont>) -> Option<(char, Vec<u8>, u32, u32, f32, f32)> {
 	
-	let Some(glyph) = font.outline_glyph(glyph) else {return None;};
+	let glyph = font.outline_glyph(glyph)?;
 	let bounds = glyph.px_bounds();
 	
 	let alpha = foreground.a as f32;